Improved the fgets function on windows, eliminating weird pauses in the console on input. It now filters "\r\n" to fix the bug in Lua's debug console where it is unable to continue.
The original implementation had an array out of bounds issue. The other issue was apparent on windows console, where the input would sometimes pause until there was an event on the console window (either another key press or a mouse movement would trigger the input to be processed). The improved fgets uses a StreamReader instead of reading directly from the Stream. I still don't understand exactly what was causing the issue, but this is now equivalent to using Console.Read() when the stream is stdin, and works as expected. I also filter "\r\n" sequences to "\n" which solves the problem of Lua's debug.debug() not returning when cont is entered into the prompt.